Add tools index page
